refactor recent home && tambah fitur category all
